Reorganize directory structure regarding data loaders
