+++ api/graphql/views.py
@@ -1,24 +1,28 @@
+import os
+from importlib import import_module
+import graphene
 from django.utils.functional import cached_property
 from graphene_django.views import GraphQLView
-from api.graphql.loaders import AuthorsByArticleIdLoader, CommentsByArticleIdLoader
 
 
-class GQLContext:
+packages = [name for name in os.listdir('./api') if
+            os.path.isdir(f'./api/{name}')
+            and name[0] != '_'
+            and os.path.exists(f'./api/{name}/context.py')]
+
+context_classes = []
+for package in packages:
+    module = import_module(f'api.{package}.context')
+    if hasattr(module, 'GQLContext'):
+        context_classes.append(module.GQLContext)
+
+class GQLContext(*context_classes):
     def __init__(self, request):
         self.request = request
 
     @cached_property
     def user(self):
         return self.request.user
-
-    @cached_property
-    def authors_by_article_id_loader(self):
-        return AuthorsByArticleIdLoader()
-
-    @cached_property
-    def comments_by_article_id_loader(self):
-        return CommentsByArticleIdLoader()
-
 
 class CustomGraphQLView(GraphQLView):
     def get_context(self, request):
