+++ api/user/context.py
@@ -1,0 +1,8 @@
+from django.utils.functional import cached_property
+from api.user.loaders import AuthorsByArticleIdLoader
+
+
+class GQLContext:
+    @cached_property
+    def authors_by_article_id_loader(self):
+        return AuthorsByArticleIdLoader()
